fix: change type total_price

+++ back-end/src/database/migrations/20220920191130-create-sale.js
@@ -29,7 +29,7 @@ module.exports = {
         },
       },
       total_price: {
-        type: Sequelize.DOUBLE(9, 2),
+        type: Sequelize.DECIMAL(9, 2),
         allowNull: false,
       },
       delivery_address: {
